sw reqif-xhtml import: add a new AllowedRTFOLEMimeTypes parameter
The HTML import is an old-style filter, so it has no XFilter
implementation where filter() would get custom parameters out of the
box. One way would be to fix by adding one more entry to the aFormalArgs
table under sfx2/, but doing that with a random parameter of a random
import filter feels dirty.

So instead make SfxMedium store all arguments as-is, this way accessing
other keys is as easy to accessing the already available FilterOptions
(string) key.

Regarding the actual filter change, don't require "text/rtf" as a mime
type for embedded objects in the reqif XHTML import, so that in case the
file has e.g. application/rtf, then that works as well.

In case an (UNO) client wants to still limit the accepted set of MIME
types, that's possible via the new parameter.

Change-Id: Ia60da44d692f550d8ad5bbf374171cac67d9e04f
Reviewed-on: https://gerrit.libreoffice.org/85229
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins