tools: define DbgTestSolarMutex() unconditionally
See the discussion at <https://gerrit.libreoffice.org/#/c/58170/>, and
this came up on IRC today again.

The above change broke the invariant that you can mix product and debug
(but not dbgutil) builds. Restore this, without mandating dbgutil for
the solar mutex assert code, which is useful for plain debug builds as
well.

Change-Id: I1f8bdb114b129fc4f39f186ba917e35e346a16b5
Reviewed-on: https://gerrit.libreoffice.org/85369
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>