sfx2: tolerate missing LOKNotifier on FloatingWindow hide
It seems the assumption is that if we get to hide state
then we must have initialized properly and set a LOKNotifier.
But this seems not to always hold, for whatever reason.

Here we are more forgiving if we don't have LOKNotifier.

Change-Id: Id21ec83489776f34228dd31f63e711b0dd77af55
Reviewed-on: https://gerrit.libreoffice.org/85401
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit 4bdc06672854a9ad121d66dc606ca50712223a80)
Reviewed-on: https://gerrit.libreoffice.org/85686
Tested-by: Jenkins
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>