use copy-on-write for SkiaSalBitmap data
E.g. scaling works by first making a copy using Create() and then
Scale() is called for the copy. This means there was a needless
data copy for mBuffer in Create(). Also make sure SkBitmap is properly
unshared if needed, as it seems copies of it always share the pixels.

Change-Id: I30a758c84e7218e9afe516477aa8429364ef8607
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/85543
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>