tdf#129522 SvxShadowItem/writerfilter: allow shadow_NONE overrides
Both editeng and writerfilter were ignoring the case where NONE
should override an inherited shadow location. Of course,
this situation is very rare. Nearly all unit tests ran when
asserting that eLocation is already NONE.

The same is true for DOC, except that it is not affected by
SvxShadowItem's PutValue problem (and neither is ODT).

RTF is not even removing the borderlines themselves,
so no point in worrying about the shadow for RTF.

Change-Id: I7c1ae67270dde81915daee2f0282aa2074d2ec8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/85658
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>