tdf#129522 ooxmlexport: allow char shadow_NONE override
Normally we would NOT want to write out nil borders or shadows.
However, if these would be inherited, NILs needs to be written out.
This already worked for paragraphs, so it was just a matter
of identifying the character styles and testing against them.

I couldn't find anything that clearly stated that ChpIter
is the CharStyle properties, but it appears to act that
way, since the first check is for ISet (which seems to be
direct properties) and the second check is usually for
styles, and the Internet also suggests Fkp's are for styles.

This is the last patch in a series, and provides the
unit test for a previous import fix.

import depends on commit 4ba8a58790ad14ea836dec47dc0ab5f47a8a55f1
which depends on commit d667eac3465b4ef0968c7341a66f86a09ca82064
and for export depends on commit
4f6dd8409e088bd2646e2856b263e5c44549cec5.

Change-Id: I5230adc52196b9b610d7863397bf33143682370c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/85870
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>