tdf#129205 DOCX import: handle the <w:shd w:val="nil" ...> paragraph property
Reading the spec, "nil" is the opposite of "clear": i.e. if the
(background) color is red and the fill (color) is green, then "clear"
means green. And you would expect "nil" means red, but it's just nothing
in Word.

Fix the problem by doing the same: don't set any paragraph property for
the "nil" case and keep doing it for the common "clear" case.

Change-Id: I30af8a7fb55fb9bab2d12e120069a479fc7ab0a9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86096
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins