Resolves: tdf#129763 resize to request happening on every activate
even though the state remains the same, so only do the resize
if the EnableClearFormatButton[All] state changes.

also change the default to off, which is what writer typically
set it to, and make calc the one that enables it explicitly. Its
easier to grow a dialog than shrink a dialog.

Change-Id: I449df7b7dcf0204b5143b8ad88e528668b84d0a3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86613
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>