tdf#129360: fix old unsigned underflow
... discovered thanks to 27926bbf3fe4bd5fee876583436d691052dd8ccb:
properties count may be 0, and unconditionally subtracting 3 from
that will give a big number. Previously this was masked by a "Very
Hot Fix" in SbxArray::GetRef that was dropped in the said commit.

This unifies property count correction in the two places that use
it: in WatchTreeListBox::RequestingChildren, where it was correct,
and in WatchTreeListBox::UpdateWatches, where underflow happened.

Change-Id: Ie980bc9571b2555cc2f643e770d824d5e2eb3731
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86700
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>