tdf#129516 DOCX import: fix incorrect z-order with textboxes
Commit 3ed23eb8b40ea33431bfb6488decb5ce0d527422 (related tdf#100074:
prepare group shapes text input via writerfilter, 2018-10-01) tweaked
the oox code, so that later it'll be able to call back to writerfilter
to parse group shape text. That makes sense, but it also removed the
reset of the group shape context, which means that two subsequent group
shapes are now imported as a single group shape with a merged child
list.

Reset the group shape context again when writerfilter asks for the
XShape from oox. If this causes a problem for the above scenario later,
then it could be considered to handle this in
ShapeContextHandler::endFastElement().

Change-Id: I14f7f0bab2c66c8430313d5b2daffe3160a58c27
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86712
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>