Revert "avoid needless 32bpp->24+8bpp conversion in png writer" (tdf#130128)
The separate alpha channel that VCL normally uses really messes up
trying to switch to 32bpp bitmaps all over the code. In the case of
tdf#130128 the bitmap has data as ScanlineFormat::N32BitTcMask,
and this uses class ColorMask, which handles only separate alpha
and drops it from the bitmap data itself. So I'm just giving up on the 32bpp
idea in Skia for somewhen later maybe.

This reverts commit c9abf6c9a8511c18312487821b738f9791d0504f.

Change-Id: Iff419e8162c4e0f0680bf62d2b078f9ea809959e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87467
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>