python3+WIN: don't fail copy of openssl DLLs+PDBs
The LO python3 target fails for me on Windows with:
  C:\lode\dev\core\workdir\UnpackedTarball\python3\PCBuild\openssl.props(24,5):
    error MSB3030: Datei "C:/lode/dev/core/workdir/UnpackedTarball/openssl/out32dll\libcrypto-1_1.dll"
    konnte nicht kopiert werden, da die Datei nicht gefunden wurde.
    [C:\lode\dev\core\workdir\UnpackedTarball\python3\PCBuild\_ssl.vcxproj]

Same for
  "C:/lode/dev/core/workdir/UnpackedTarball/openssl/out32dll\libssl-1_1.pdb"
  "C:/lode/dev/core/workdir/UnpackedTarball/openssl/out32dll\libcrypto-1_1.pdb"
  "C:/lode/dev/core/workdir/UnpackedTarball/openssl/out32dll\libssl-1_1.dll"

Other files were also renamed in a previous hunk of this patch.

For other people these failures are silently ignored, but they
show up in their python3 build.log. But my msbuild version
15.9.21+g9802d43bc3 from VS2017 fails hard on these errors.

So this just adapt the pdb and dll names to match the previous
renames, which passes the copy calls, so the build continues.

Change-Id: Iffc848c334caec534f6e99f8bf83a42da570bbb1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87358
Reviewed-by: Michael Stahl <michael.stahl@cib.de>
Tested-by: Jenkins