tdf#103964 DOCX import: ignore rotation when setting position of group shapes
Regression from commit 3e5ec561a36e08c7b6f0ac1e097330539ab6cc1c (DOCX
import: make sure rotation does not affect shape position, 2014-09-24),
the motivation for tweaking the rotation when setting the position of a
shape was for images. By accident, this also happened for group shapes.

But the bugdoc shows it's not a good idea to read the rotation of group
shapes: the group shape is just a container, it does not have rotation
in itself.

(The test120551 intention was probably just to verify that the position
is not entirely off, so the small required change to the expected value
should be OK.)

Change-Id: I8e12c28e65c5f64168c3f802546fddf472fcc6eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87551
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>