Use the "Double-tap" term instead of "click" on all iOS devices (also iPad)
In general, the term "mobile," when used in APIs related to Online and
the mobile apps, means "mobile phone but not tablet". Which is sad,
but it is too late to change that convention now.

Thus comphelper::LibreOfficeKit::isMobile() returns false on iPad,
apparently intentionally. But surely we want to use the "double tap"
term instead of "click" also on iPad.

Change-Id: I8f7186ea6590f2a21b885549751346edad09eb8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87858
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
(cherry picked from commit 31d3f1b3948df707183e10120278b9c2b1410bc9)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87931
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>