There should be no need to mark nested PropertyPairHash as SVXCORE_DLLPUBLIC
For one, the surrounding class SdrCustomShapeGeometryItem is already marked
SVXCORE_DLLPUBLIC (whether or not that impacts the default export status of the
nested struct on the various platforms, though).  And, for another, the struct
only has inline members (which should make the previous point moot).

Also, to be on the safe side, mark the PropertyPairHash member function as
inline already on its first declaration.

(This is a follow-up to 67b8e5e4b4546f487be569b7da103f5298645bd5 "tdf#128302:
Split SVXCORE_DLLPUBLIC from SVX_DLLPUBLIC".)

Change-Id: I6716a589060d816ca822583711d0e559ea114fb7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87865
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>