deb#949754: unotools,officecfg: don't ODF export using ODFVER_UNKNOWN
The problem was that commit 564a9a371b0a989872f969dc3c67f3accbeeed58
temporarily added ODF 1.3 version strings to the optsavepage.ui,
just to get translations for them; unfortunately the dialog itself
was not adapted to the new values, so when you select them the
result is 0, or ODFVER_UNKNOWN... fortunately this was reverted
before 6.4.0.3.

The value ODFVER_UNKNOWN is very dubious and without an obvious purpose;
http://specs.openoffice.org/appwide/odf/odf_1-2_migration.odt
mentions it but provides no details.

Hence let's interpret it the same as ODFVER_LATEST for export purposes.

Also add the value ODFVER_012_EXT_COMPAT to the configuration, because
this is actually used.

Also fix a copypasta in SvtSaveOptions_Impl::IsReadOnly(), where the
wrong flag is checked for OdfDefaultVersion.

Change-Id: Ie276f0ef3cead3ffa016ba939aede74581fb0257
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87900
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>
Reviewed-by: Michael Stahl <michael.stahl@cib.de>