tdf#129912 correctly stop unstyled footnote parsing
The bug document somehow manages to generated a footnote, which
never terminates the format loop in SwTextFrame::Format_.

It contains an unstyled footnote, which I wasn't able to reproduce
to create in Word. So I manually edited the XML of the included
unit test document, which I used to develop the original patch,
and which reproduces the broken parsing behaviour.

This patch correctly stops the parsing of the custom footnote
reference, if the text run containing the footnote reference is
finished, which also fixes loading the bug document.

The unit test checks various footnote variants, which represent
different problems I found when developing the custom footnote
parsing in commit 0a093972f4bce06984684f3460e73511fbf5f8c8
("tdf#121441 improve DOCX footnote import") and now also includes
an unstyled one.

It also contains a (still?) broken footnote test, with a complex
differing footnote.

Change-Id: I748955285d76b6f3122d1da5d8823068f3d7633f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87981
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>