tdf#130362 sw: fix anchoring of inline math objects
Regression from 8641c42c5a341536cc6f4509305a12d511e51856 (sw: insert
image: set anchor to at-char by default, 2019-11-18), that defaulted to
at-char anchoring for charts and images.

What was not considered is that math objects had a previous as-char
default (not to-para), and that is supposed to be unchanged.

Change-Id: I2a91af6425035b48a0e47ad9b10939945855cd16
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87976
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins