Classify shape(s) selections as 'complex'...
in isComplex() of XTransferable2 implementation SwTransferable.
This method is used by online via doc_getSelectionType() to determine the
type of selection. In writer this method used to classify shape objects as
non-complex which translates to 'text' in online, so if a shape is selected
in online, the js code there would treat it as text. This is problematic
when you want to send the correct align uno commands based on selection type.
So returning true in isComplex() for writer would correctly treat the shape
selections as 'complex' in online Writer.

Change-Id: I09fcd9e4fab48aa0d7e7d04c88bae9e1281a1b0c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88158
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>
(cherry picked from commit ba45e91de8874b1ff8f97ec5c612acdd9a84b49c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88495
Tested-by: Jenkins
Reviewed-by: Andras Timar <andras.timar@collabora.com>