DOCX export: fix handling of section starts that originally had page margins
This is similar to commit 60eaf95d137e19b50e89819216ef0f2a4e8e8318 (DOCX
export: fix handling of section starts that originally had headers,
2020-02-06), except here the top margin has to taken from that follow
page style, not the header.

Without this, it can happen that the page number in the original Writer
doc model and the exported Word result do not match.

This required reworking WriteNextStyleHeaderFooter(), which assumed that
the header/footer status is already calculated by the time its called.
But the page margin code runs earlier, so we need to make that decision
earlier, even when the header/footer status is not yet calculated.

Change-Id: Ife7396603702d2048d544aa46f96acfa337a041a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88211
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins