tdf#130325 Fix listbox used as dropdown
In this case, the listbox is not used as a static widget,
also not as a combobox with dropdown.
Instead the listbox is placed in a popup to choose values from.
Need to handle this case similiar to the combobox dropdown
(Cursor movement only travels through items, Return selects one item).

Regression from 593454f9f84dd9dd0c126ff56fc73fa8842062fe

Change-Id: Idadc3da5847e12e1408203b13ab59eb53fe14eea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88581
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>